requirements.txt
accelerate==1.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
async-timeout==5.0.1
attrdict==2.0.1
attrs==25.4.0
babel==2.17.0
bce-python-sdk==0.9.55
bcrypt==3.2.2
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.2
cssselect==1.3.0
cssutils==2.11.1
cycler==0.12.1
Cython==3.2.2
datasets==4.5.0
dill==0.4.0
et_xmlfile==2.0.0
exceptiongroup==1.3.1
fastapi==0.123.9
filelock==3.20.0
fire==0.7.1
Flask==3.1.2
flask-babel==4.0.0
fonttools==4.61.0
frozenlist==1.8.0
fsspec==2025.10.0
future==1.0.0
gitdb==4.0.12
GitPython==3.1.45
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.41.1
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.3
googleapis-common-protos==1.72.0
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
huggingface-hub==0.36.0
hyperframe==6.1.0
idna==3.11
ImageIO==2.37.2
imgaug==0.4.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
lazy_loader==0.4
Levenshtein==0.27.3
lmdb==1.7.5
lxml==6.0.2
MarkupSafe==3.0.3
matplotlib==3.10.7
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.18
narwhals==2.13.0
networkx==3.4.2
numpy==2.2.6
oauthlib==3.3.1
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
packaging==25.0
paddleocr==2.7.0.3
pandas==2.3.3
passlib==1.7.4
pdf2docx==0.5.8
pdf2image==1.17.0
pillow==12.0.0
portalocker==3.2.0
premailer==3.10.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.1
psutil==7.1.3
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.4.0
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
PyJWT==2.10.1
PyMuPDF==1.20.2
pyparsing==3.2.5
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-Levenshtein==0.27.3
python-multipart==0.0.20
pytz==2025.2
pywin32==311
PyYAML==6.0.3
qdrant-client==1.16.1
RapidFuzz==3.14.3
rarfile==4.2
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.30.0
rsa==4.9.1
safetensors==0.7.0
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
sentence-transformers==2.7.0
shapely==2.1.2
six==1.17.0
smmap==5.0.2
soupsieve==2.8
starlette==0.50.0
streamlit==1.52.0
sympy==1.14.0
tenacity==9.1.2
termcolor==3.2.0
threadpoolctl==3.6.0
tifffile==2025.5.10
tokenizers==0.22.1
toml==0.10.2
torch==2.10.0
torchvision==0.25.0
tornado==6.5.2
tqdm==4.67.1
transformers==4.57.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.38.0
visualdl==2.5.3
watchdog==6.0.0
Werkzeug==3.1.4
xxhash==3.6.0
yarl==1.22.0
